agent/requirements.txt
livekit-agents
livekit-plugins-groq
livekit-plugins-deepgram

livekit-plugins-silero
python-dotenv
livekit-api
